# atlas_openapi_v1.yaml
openapi: 3.0.3
info:
  title: "Project: ATLAS - AI Coaching Platform API (V1)"
  description: "학생의 4축 잠재 공간 모델을 기반으로 코칭 활동을 지원하는 통합 API"
  version: "1.0.0"
servers:
  - url: "https://api.atlas.sigongjoa.com/v1"
    description: "Production Server"
  - url: "http://localhost:8000/v1"
    description: "Local Development Server"
paths:
...
  /users/me:
    get:
      summary: Get current user's profile
      operationId: get_current_user_profile
      responses:
        "200":
          description: Current user's profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMe'
        "401":
          description: Unauthorized
      tags:
        - Users
    put:
      summary: Update current user's profile
      operationId: update_current_user_profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        "200":
          description: User profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMe'
        "401":
          description: Unauthorized
      tags:
        - Users
    delete:
      summary: Deactivate current user's account
      operationId: deactivate_current_user_account
      responses:
        "200":
          description: Account deactivated successfully
        "401":
          description: Unauthorized
      tags:
        - Users
  /users/me/password:
    put:
      summary: Update current user's password
      operationId: update_current_user_password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPasswordUpdate'
      responses:
        "200":
          description: Password updated successfully
        "401":
          description: Unauthorized
        "400":
          description: Invalid current password or new passwords do not match
      tags:
        - Users
  /users/me/notifications:
    put:
      summary: Update current user's notification preferences
      operationId: update_current_user_notifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserNotificationUpdate'
      responses:
        "200":
          description: Notification preferences updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserNotificationUpdate'
        "401":
          description: Unauthorized
      tags:
        - Users
...
components:
  schemas:
...
    UserMe:
      title: UserMe
      required:
        - user_id
        - username
        - email
        - user_type
      type: object
      properties:
        user_id:
          title: User Id
          type: string
        username:
          title: Username
          type: string
        email:
          title: Email
          type: string
        user_type:
          title: User Type
          type: string
          enum:
            - student
            - coach
    UserUpdate:
      title: UserUpdate
      type: object
      properties:
        username:
          title: Username
          type: string
        email:
          title: Email
          type: string
    UserPasswordUpdate:
      title: UserPasswordUpdate
      required:
        - current_password
        - new_password
        - confirm_new_password
      type: object
      properties:
        current_password:
          title: Current Password
          type: string
        new_password:
          title: New Password
          type: string
        confirm_new_password:
          title: Confirm New Password
          type: string
    UserNotificationUpdate:
      title: UserNotificationUpdate
      type: object
      properties:
        new_assignments:
          title: New Assignments
          type: boolean
        feedback_from_coach:
          title: Feedback From Coach
          type: boolean
        platform_updates:
          title: Platform Updates
          type: boolean