# backend/schemas.py
from pydantic import BaseModel, Field, ConfigDict
from typing import List, Optional, Literal
from datetime import datetime

class StudentCreate(BaseModel):
    student_id: str
    student_name: str

class CoachCreate(BaseModel):
    coach_id: str
    coach_name: str

class ParentCreate(BaseModel):
    parent_name: str
    kakao_user_id: Optional[str] = None

class VectorHistoryEntry(BaseModel):
    vector_id: str
    assessment_id: str
    student_id: str
    created_at: datetime
    axis1_geo: int = Field(..., ge=0, le=100)
    axis1_alg: int = Field(..., ge=0, le=100)
    axis1_ana: int = Field(..., ge=0, le=100)
    axis2_opt: int = Field(..., ge=0, le=100)
    axis2_piv: int = Field(..., ge=0, le=100)
    axis2_dia: int = Field(..., ge=0, le=100)
    axis3_con: int = Field(..., ge=0, le=100)
    axis3_pro: int = Field(..., ge=0, le=100)
    axis3_ret: int = Field(..., ge=0, le=100)
    axis4_acc: int = Field(..., ge=0, le=100)
    axis4_gri: int = Field(..., ge=0, le=100)

class AssessmentCreate(BaseModel):
    student_id: str
    assessment_type: str
    source_ref_id: Optional[str] = None
    notes: Optional[str] = None
    vector_data: dict = Field(
        ...,
        example={
            "axis1_geo": 70, "axis1_alg": 60, "axis1_ana": 75,
            "axis2_opt": 80, "axis2_piv": 65, "axis2_dia": 70,
            "axis3_con": 85, "axis3_pro": 70, "axis3_ret": 75,
            "axis4_acc": 90, "axis4_gri": 80
        }
    )

class SubmissionCreate(BaseModel):
    student_id: str
    problem_text: str

class SubmissionResult(BaseModel):
    submission_id: str
    status: str
    logical_path_text: str
    concept_id: Optional[str] = None
    manim_content_url: Optional[str] = None
    submitted_at: Optional[datetime] = None

class CoachMemoCreate(BaseModel):
    coach_id: str
    student_id: str
    memo_text: str

class CoachMemoResponse(CoachMemoCreate):
    memo_id: str
    created_at: datetime

    class Config:
        from_attributes = True

class UserLogin(BaseModel):
    email_or_username: str
    password: str
    user_type: Literal["student", "coach"]

class Token(BaseModel):
    access_token: str
    token_type: str

class Notification(BaseModel):
    notification_id: str
    user_id: str
    type: Literal["assignment_graded", "new_feedback", "new_student", "new_submission", "system_alert"]
    title: str
    message: str
    created_at: datetime
    is_read: bool
    related_id: Optional[str] = None

class UserMe(BaseModel):
    user_id: str
    username: str
    email: str
    user_type: Literal["student", "coach"]

class UserUpdate(BaseModel):
    username: Optional[str] = None
    email: Optional[str] = None

class UserPasswordUpdate(BaseModel):
    current_password: str
    new_password: str
    confirm_new_password: str

class UserNotificationUpdate(BaseModel):
    new_assignments: bool
    feedback_from_coach: bool
    platform_updates: bool

class WeeklyReport(BaseModel):
    report_id: int
    student_id: str
    coach_id: Optional[str] = None
    period_start: datetime
    period_end: datetime
    status: str
    ai_summary: str
    vector_start_id: str
    vector_end_id: str
    coach_comment: Optional[str] = None
    created_at: datetime
    finalized_at: Optional[datetime] = None

class LLMFeedback(BaseModel):
    log_id: Optional[int] = None
    coach_feedback: str
    reason_code: Optional[str] = None
    source_submission_id: Optional[str] = None

class LLMLogResponse(BaseModel):
    log_id: int
    source_submission_id: Optional[str] = None
    decision: str
    model_version: Optional[str] = None
    coach_feedback: Optional[str] = None
    reason_code: Optional[str] = None
    created_at: datetime
    model_config = ConfigDict(from_attributes=True)

class CoachComment(BaseModel):
    coach_comment: str


class Student(BaseModel):
    student_id: str
    student_name: str
    created_at: datetime
    model_config = ConfigDict(from_attributes=True)

class Coach(BaseModel):
    coach_id: str
    coach_name: str
    model_config = ConfigDict(from_attributes=True)

class Parent(BaseModel):
    parent_id: int
    parent_name: str
    kakao_user_id: Optional[str] = None
    model_config = ConfigDict(from_attributes=True)

class Curriculum(BaseModel):
    curriculum_id: str
    curriculum_name: str
    description: Optional[str] = None
    model_config = ConfigDict(from_attributes=True)

class Concept(BaseModel):
    concept_id: str
    curriculum_id: Optional[str] = None
    concept_name: str
    manim_data_path: Optional[str] = None
    description: Optional[str] = None
    model_config = ConfigDict(from_attributes=True)

class ConceptRelation(BaseModel):
    relation_id: int
    from_concept_id: str
    to_concept_id: str
    relation_type: str
    model_config = ConfigDict(from_attributes=True)

class StudentMastery(BaseModel):
    student_id: str
    concept_id: str
    mastery_score: int
    status: Optional[str] = None
    last_updated: Optional[datetime] = None
    model_config = ConfigDict(from_attributes=True)

class AnkiCard(BaseModel):
    card_id: int
    student_id: str
    llm_log_id: int
    question: str
    answer: str
    next_review_date: datetime
    interval_days: int
    ease_factor: float
    repetitions: int
    model_config = ConfigDict(from_attributes=True)

class AnkiCardReview(BaseModel):
    grade: int = Field(..., ge=0, le=5)